refactor: Cambiar carga de Fuentes y ajuste texto Actividades
Resuelve: udistrital/financiera_documentacion#411
